% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data frame that contains the edges of the graph and the weight of the edge}

\item{init_node}{A numeric scalar that exists in the graph}
}
\value{
The shortest distance from \code{init_node} to every other node
}
\description{
Dijkstra's algorithm
}
\examples{
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)
}
\references{

}
