% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{An integer}

\item{b}{An integer}
}
\value{
The greatest common divisor of \code{a} and \code{b}
}
\description{
Euclidean algorithm
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
