% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data frame that contains the edges of the graph and the weight of the edge}

\item{init_node}{A numeric scalar that exists in the graph}
}
\value{
The shortest distance from \code{init_node} to every other node
}
\description{
Dijkstra's algorithm
}
\examples{
dijkstra(data.frame(v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6), v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5), w=c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9)), 1)
dijkstra(data.frame(v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6), v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5), w=c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9)), 3)
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstras_algorithm/}
}
