% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{An example of a weighted graph from Wikipedia.}
\format{
A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{Start of the edge}
  \item{v2}{End of the edge}
  \item{w}{Weight of the edge}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstras_algorithm/}
}
\usage{
wiki_graph
}
\description{
A weighted graph containing 6 nodes.
}
\keyword{datasets}
